+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - details dict in remote mode search now contains 'date' field
 - fix tests relying on OpenAlyx
 - warning instead of error raised when tag_mismatched_dataset REST query returns 403
+- list_datasets and ses2records now better handle sessions with no datasets in remote mode
 
 ## [1.9.1]